requirements.txt
diffusers==0.27.2
transformers==4.39.3
accelerate==0.28.0
sentencepiece
gradio==4.29.0
einops
safetensors
pillow
opencv-python
clip-anytorch
torchmetrics
tqdm
